requirements.txt
Django~=4.2.16
psycopg2==2.9.9
Jinja2==3.1.3
social-auth-app-django==5.4.0
django-environ~=0.11.2
djangorestframework~=3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
drf-yasg~=1.21.8
requests~=2.32.3
social-auth-core~=4.5.4
isort~=6.0.0
flake8~=7.1.2
